Add project based fastai2 integration tests
