+++ .github/workflows/ci.yml
@@ -117,8 +117,10 @@ jobs:
           python-version: 3.7
       - name: Install test dependencies
         run: ./ci/install_test_deps.sh
+      - name: Install fastai
+        run: python -m pip install fastai
       - name: Run tests
-        run: ./ci/fastai_integration_tests.sh
+        run: ./ci/test_project.sh tests/integration/projects/fastai2
       - name: Upload test coverage to Codecov
         uses: codecov/codecov-action@v1.0.12
 
@@ -141,7 +143,7 @@ jobs:
         uses: codecov/codecov-action@v1.0.12
 
   api_server_integration_tests:
-    name: ${{ matrix.os }} API Server Integration Tests
+    name: API Server Integration Tests (${{ matrix.os }})
     runs-on: ${{ matrix.os }}
     strategy:
       fail-fast: false
@@ -158,12 +160,14 @@ jobs:
       - name: Install test dependencies
         run: ./ci/install_test_deps.sh
       - name: Run tests
-        run: ./ci/integration_tests.sh
+        run: ./ci/test_project.sh tests/integration/projects/general
+      - name: Run tests
+        run: ./ci/test_project.sh tests/integration/projects/general_non_batch
       - name: Upload test coverage to Codecov
         uses: codecov/codecov-action@v1.0.12
 
   back_compatibility_integration_tests:
-    name: ${{ matrix.os }} Back Compatibility Integration Tests
+    name: Backward Compatibility Integration Tests (${{ matrix.os }})
     runs-on: ${{ matrix.os }}
     strategy:
       fail-fast: false
