--- tests/integration/fastai_utils.py
@@ -1,5 +1,0 @@
-import numpy as np
-
-
-def get_items(x):
-    return np.ones([5, 5], np.float32)
