+++ tests/integration/projects/conftest.py
@@ -5,7 +5,6 @@ import os
 import sys
 
 import pytest
-
 from tests.integration.utils import (
     build_api_server_docker_image,
     run_api_server,
@@ -35,7 +34,6 @@ def host(pytestconfig, clean_context, enable_microbatch):
     test_svc_bundle = pytestconfig.getoption("bento_dist") or os.path.join(
         sys.argv[1], "build", "dist"
     )
-    print(test_svc_bundle)
 
     if pytestconfig.getoption("docker"):
         image = clean_context.enter_context(
@@ -46,3 +44,14 @@ def host(pytestconfig, clean_context, enable_microbatch):
     else:
         with run_api_server(test_svc_bundle, enable_microbatch) as host:
             yield host
+
+
+@pytest.fixture(scope="module")
+def service(pytestconfig):
+    test_svc_bundle = pytestconfig.getoption("bento_dist") or os.path.join(
+        sys.argv[1], "build", "dist"
+    )
+
+    import bentoml
+
+    return bentoml.load_from_dir(test_svc_bundle)
