+++ tests/integration/projects/fastai2/service.py
@@ -1,5 +1,9 @@
-import bentoml
+import pathlib
+import sys
+
 import numpy as np
+
+import bentoml
 from bentoml.adapters import DataframeInput
 from bentoml.frameworks.fastai import FastaiModelArtifact
 
@@ -13,3 +17,16 @@ class FastaiClassifier(bentoml.BentoService):
         _, _, output = self.artifacts.model.predict(input_data)
 
         return output.squeeze().item()
+
+
+if __name__ == "__main__":
+    artifacts_path = sys.argv[1]
+    bento_dist_path = sys.argv[2]
+    service = FastaiClassifier()
+
+    from model.model import Loss, Model  # noqa # pylint: disable=unused-import
+
+    service.artifacts.load_all(artifacts_path)
+
+    pathlib.Path(bento_dist_path).mkdir(parents=True, exist_ok=True)
+    service.save_to_dir(bento_dist_path)
